:recycle: Refactor posts, experiences and projects files
