+++ public/experiences.json
@@ -5,7 +5,17 @@
       "company": "Contta Inteligência Fiscal",
       "role": "Front-End Developer",
       "work_time": "Sept. 2020 - Nov. 2021",
-      "description": "Developed and maintained the main company's project, which was a Dashboard containing financial information of all registered companies by users."
+      "description": "Developed and maintained the main company's project, which was a Dashboard containing financial information of all registered companies by users.",
+      "technologies": [
+        {
+          "id": 1,
+          "name": "React"
+        },
+        {
+          "id": 2,
+          "name": "Rest API"
+        }
+      ]
     }
   ],
   "ptBR": [
@@ -14,7 +24,17 @@
       "company": "Contta Inteligência Fiscal",
       "role": "Desenvolvedor Front-End",
       "work_time": "Set. 2020 - Nov. 2021",
-      "description": "Desenvolveu e deu manutenção no principal projeto da empresa: um Dashboard que contém informações financeiras de todas as empresas cadastradas pelo usuário."
+      "description": "Desenvolveu e deu manutenção no principal projeto da empresa: um Dashboard que contém informações financeiras de todas as empresas cadastradas pelo usuário.",
+      "technologies": [
+        {
+          "id": 1,
+          "name": "React"
+        },
+        {
+          "id": 2,
+          "name": "Rest API"
+        }
+      ]
     }
   ]
 }
