+++ src/modules/common/lib/projects.ts
@@ -1,21 +1,10 @@
 import fs from 'fs';
 import path from 'path';
 
-type Locale = 'en' | 'ptBR';
-
-type Project = {
-  id: number;
-  company: string;
-  role: string;
-  work_time: string;
-  description: string;
-};
-
-type ProjectsByLocale = Record<Locale, Project[]>;
-
-type GetProjectsParams = {
-  locale?: Locale;
-};
+import type {
+  GetProjectsParams,
+  ProjectsByLocale,
+} from '@/modules/common/types';
 
 export function getProjects(
   { locale = 'en' }: GetProjectsParams = { locale: 'en' },
@@ -24,10 +13,6 @@ export function getProjects(
   const projectFile = fs.readFileSync(projectFilePath, 'utf-8');
 
   const parsedProjects = JSON.parse(projectFile) as ProjectsByLocale;
-  const parsedProjectsWithImage = parsedProjects[locale].map(project => ({
-    ...project,
-    image: `/images/projects/${project.id}.jpg`,
-  }));
 
-  return parsedProjectsWithImage;
+  return parsedProjects[locale];
 }
